Fix Average Vote being an Integer
The Average Vote value was previously queried as an integer
and passed to number_format($avg, 2) afterwards, resulting in X.00
every time. This fix updates the average count to properly return
a value with two decimals.

Bug: T279040

+++ includes/Vote.php
@@ -69,7 +69,7 @@ class Vote {
 				$dbr = wfGetDB( DB_REPLICA );
 				$setOpts += Database::getCacheSetOptions( $dbr );
 
-				return (int)$dbr->selectField(
+				return (float)$dbr->selectField(
 					'Vote',
 					'AVG(vote_value)',
 					[ 'vote_page_id' => $this->PageID ],
